Moved Context Extras nodes to its own file
